src/qubo/requirements.txt
amplify==1.4.1
python-dotenv==1.2.1
